apps/api/requirements.txt
# Django Backend - Requirements
# Pin versions for stability

# Django core
Django==4.2.8
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==23.5

# Storage
boto3==1.34.11
django-storages==1.14.2
minio==7.2.0

# Celery
celery==5.3.4
redis==5.0.1

# Image processing
Pillow==10.1.0

# Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.27

# Development tools
black==23.12.1
ruff==0.1.9
isort==5.13.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
